requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
filterpy==1.4.5
fonttools==4.39.3
idna==3.4
imageio==2.28.1
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-loader==0.2
lit==16.0.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mysql-connector-python==8.0.33
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
schedule==1.2.0
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.22.1
six==1.16.0
sympy==1.11.1
thop==0.1.1.post2209072238
tifffile==2023.4.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.93
urllib3==1.26.15
zipp==3.15.0
opencv-python==4.7.0.72
paho-mqtt==1.6.1
